Weeded out implicit narrowing casts in src/main/[a-z]*.cpp, as revealed by -Wconversion.
git-svn-id: http://projects.cs.kent.ac.uk/projects/cxxr/svn/trunk@1468 549f2c98-2231-0410-b618-a09cd5e121c0

+++ src/main/connections.cpp
@@ -5835,7 +5835,7 @@ do_memDecompress(SEXP call, SEXP op, SEXP args, SEXP env)
 		break;
 	    }
 	}
-	outlen = strm.total_out;
+	outlen = size_t(strm.total_out);
 	lzma_end(&strm);
 	ans = allocVector(RAWSXP, outlen);
 	memcpy(RAW(ans), buf, outlen);
